Separate API and weather data collection #12
Separating API and weather data collection into two processes. This will allow us to independently
scale each if we need to do so. Also, if one goes down, the other is still up and running. Also
adding configs as yaml so we can easily reference and set things in the future